memory 64 decode and validation

+++ ast/src/commonMain/kotlin/io/github/charlietap/chasm/ast/instruction/MemArg.kt
@@ -1,3 +1,3 @@
 package io.github.charlietap.chasm.ast.instruction
 
-data class MemArg(val align: UInt, val offset: UInt)
+data class MemArg(val align: UInt, val offset: ULong)
