+++ executor/instantiator/src/commonMain/kotlin/io/github/charlietap/chasm/executor/instantiator/allocation/memory/MemoryAllocator.kt
@@ -23,7 +23,7 @@ internal inline fun MemoryAllocator(
     type: MemoryType,
     crossinline memoryFactory: LinearMemoryFactory,
 ): Address.Memory {
-    val pages = LinearMemory.Pages(type.limits.min)
+    val pages = LinearMemory.Pages(type.limits.min.toUInt())
     val memory = memoryFactory(pages)
     val instance = MemoryInstance(type, memory)
