+++ executor/instantiator/src/commonTest/kotlin/io/github/charlietap/chasm/executor/instantiator/runtime/allocation/memory/MemoryAllocatorTest.kt
@@ -24,7 +24,7 @@ class MemoryAllocatorTest {
         )
 
         val min = 3
-        val limits = limits(min.toUInt())
+        val limits = limits(min.toULong())
         val type = memoryType(limits = limits)
 
         val memory = linearMemory()
