+++ ir-factory/src/commonMain/kotlin/io/github/charlietap/chasm/ir/factory/MemArgFactory.kt
@@ -7,5 +7,5 @@ internal inline fun MemArgFactory(
     arg: MemArg,
 ) = IRMemArg(
     align = arg.align,
-    offset = arg.offset,
+    offset = arg.offset.toUInt(),
 )
