+++ test/fixture/type-system/src/commonMain/kotlin/io/github/charlietap/chasm/fixture/type/Limits.kt
@@ -3,8 +3,8 @@ package io.github.charlietap.chasm.fixture.type
 import io.github.charlietap.chasm.type.Limits
 
 fun limits(
-    min: UInt = 0u,
-    max: UInt? = null,
+    min: ULong = 0u,
+    max: ULong? = null,
 ) = Limits(
     min = min,
     max = max,
